Use different function for timestamps
When processing logs `current_time()` gives wrong timestamps (i.e., the
time stamps are mostly the same for all events).  Using
`current_event_time()` fixes that.

The problem is that when capturing real network data it's the other way
around — the timestamps are mostly the same when using
`current_event_time()`.

Deploying the original version with `current_time()` reveals that the
problem is there, so `current_event_time()` improves the situation at
least for pcaps.