Inline COI bitfield
The problem is that `coi` name is part of the bitfield so parent object
cannot use the same name since the bitfield is anonymous.  If we choose
to use a `CoI` unit then we need to change all other structures.

It is only used in one Information Object (M_EI_NA_1) so it's easier to
introduce some inconsistency back.