Fix C_RD_NA_1 parsing
Have not seen this ASDU in any of our traffic, but pcaps in
`testing/Traces/third` have these (seems that this ASDU type is used as
an attack vector since the pcaps are all about C_RD_NA_1).

This ASDU has variable length (depending on PN flag in Cause of
Transmision field).  Missed this bit since Beckhoff documentation
specifies that there is no object data.

Had to work around a Spicy "quirk", described in a
comment in the code.